SI-51 Create TestFlightConfigStorage subclass
We're going to reuse these changes for the "Add Tester" slash command
