+++ botto/storage/__init__.py
@@ -3,6 +3,7 @@ from . import reminder_storage
 from . import timezone_storage
 from . import enablement_storage
 from . import config_storage
+from . import testflight_config_storage
 from .beta_testers import beta_testers_storage
 
 MealStorage = meal_storage.MealStorage
@@ -11,4 +12,5 @@ ReminderStorage = reminder_storage.ReminderStorage
 TimezoneStorage = timezone_storage.TimezoneStorage
 EnablementStorage = enablement_storage.EnablementStorage
 ConfigStorage = config_storage.ConfigStorage
+TestFlightConfigStorage = testflight_config_storage.TestFlightConfigStorage
 BetaTestersStorage = beta_testers_storage.BetaTestersStorage
